+++ test/run.sh
@@ -20,3 +20,4 @@ shellcheck provision/*.sh
 bats test/*.bats
 bats test/include/*.bats
 bats test/provision/*.bats
+bats test/contrib/*.bats